+++ Root/src/PAPERO_info.cpp
@@ -63,7 +63,7 @@ int main(int argc, char *argv[])
     foutput->cd();
 
     // Find if there is an offset before first event
-    unsigned int offset = 0;
+    uint64_t offset = 0;
     offset = seek_first_evt_header(file, offset, verbose);
     int padding_offset = 0;
